--- a/.gitignore
+++ b/.gitignore
@@ -140,3 +140,4 @@ data2/
 /ignore/
 data-moved/
 all-puzzle-answers.txt
+newdata*