--- a/actorle_solver.py
+++ b/actorle_solver.py
@@ -2,6 +2,7 @@ import argparse
 import collections
 
 import pandas as pd
+import numpy as np
 
 from cli import SmartFormatter
 from movie_clues import write_movie_clues_file, read_puzzle_clues
@@ -90,6 +91,7 @@ def make_movie_title_word_regex(movie_title_word):
 
 def get_matching_movie_ids(titles_data_frame, movie_clue):
     titles_data_frame = titles_data_frame[titles_data_frame.startYear == movie_clue.year]
+    titles_data_frame = titles_data_frame[titles_data_frame.averageRating.isin([movie_clue.score])]
     print("Filtered down to {} movies from the year {}".format(titles_data_frame.shape[0], movie_clue.year))
     match_pattern = make_movie_title_regex(movie_clue.title_pattern)
     query = "primaryTitle.str.match('{}')".format(match_pattern)