--- a/clues-files/actorle-2022-jul-31.txt
+++ b/clues-files/actorle-2022-jul-31.txt
@@ -1,9 +1,16 @@
-xxx xxxxxxx|1986|Adventure,Drama,Action|0.0
-xxxx xxxxxxx|1988|Thriller,Horror|0.0
-xxxxxxxx xx xxxxxxx|1990|Drama|0.0
-xxx xxxx: xxxx x xxxxxxxxx|1995|Action,Thriller|0.0
-xxx xxx xx xxx xxxx xxxx|1998|Adventure,Action,Drama|0.0
-xxxxxxxx & xxxxxxx|2000|Drama,Adventure,Fantasy|0.0
-xxxxxxxx|2005|Adventure,Comedy,Drama|0.0
-xxxxxx|2006|Fantasy,Action,Adventure|0.0
-xxxxxxxxx|2008|Drama,Western,Crime|0.0
+xxx xxxxxxx|1986|Adventure,Drama,Action|7.4
+xxxx xxxxxxx|1988|Thriller,Horror|7.2
+xxxxxxxx xx xxxxxxx|1990|Drama|7.2
+xxx xxxx: xxxx x xxxxxxxxx|1995|Action,Thriller|7.6
+xxx xxx xx xxx xxxx xxxx|1998|Adventure,Action,Drama|6.5
+xxxxxxxx & xxxxxxx|2000|Drama,Adventure,Fantasy|3.6
+xxxxxxxx|2005|Adventure,Comedy,Drama|6.5
+xxxxxx|2006|Fantasy,Action,Adventure|5.1
+xxxxxxxxx|2008|Drama,Western,Crime|6.7
+xxx xxxx xxxxxxx x|2009|Comedy,Mystery|5.6
+xxxxxx xxxx|2011|Thriller,Drama|7.1
+xxx xxxxx|2012|Drama,Thriller|7.0
+xxxxxx x xxxxxxxx: xxxx xx xxxxxxx|2016|Action,Adventure,Fantasy|6.4
+xxx xxx xxx xxxx xxxxxxxx|2016|Drama|7.2
+xxxxxxxx'x xxxxx|2016|Drama|5.6
+xxxxx xx xxxxx|2021|Drama,Crime,History|6.6