--- a/tests/integration/accuracy-test.py
+++ b/tests/integration/accuracy-test.py
@@ -21,7 +21,7 @@ def read_expected_answers(answers_file_path):
 def solve_puzzle(clues_file_path, solver_script_path):
     shell_cmd = (
         "bash {} --clues-file {} | "
-        "grep -i dude | "
+        "grep \"Dude - I think\" | "
         "awk -F \"Dude - I think it's... \" '{{print $NF}}' | "
         "awk -F '\\!' '{{print $1}}'".format(solver_script_path, clues_file_path))
     return subprocess.getoutput(shell_cmd)