- Fixes #4 
- Fixes #3 

## Before
```
time PYTHONPATH=$PYTHONPATH:. python tests/integration/accuracy-test.py -p clues-files -a data/puzzle-actual-answers.csv -s /Users/mickyfitz/workspace/knacktorle/solve-current-actorle-puzzle.sh

                                         Summary
┏━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━┳━━━━━━━━━━━━━┓
┃ Puzzle                  ┃ Expected Answer ┃ Actual Answer       ┃ Result ┃ Time        ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━╇━━━━━━━━━━━━━┩
│ actorle-2023-jan-10.txt │ ********        │ ********            │ PASS   │ 0:00:02.926 │
│ actorle-2023-jan-04.txt │ ********        │ ********            │ PASS   │ 0:00:02.764 │
│ actorle-2022-dec-15.txt │ ********        │ ********            │ PASS   │ 0:00:02.978 │
│ actorle-2023-mar-07.txt │ Ansel Elgort    │ Soumitra Chatterjee │ FAIL   │ 0:00:02.509 │
│ actorle-2022-apr-04.txt │ ********        │ ********            │ PASS   │ 0:00:02.872 │
│ actorle-2022-apr-28.txt │ ********        │ ********            │ PASS   │ 0:00:02.822 │
│ actorle-2023-jan-17.txt │ ********        │ ********            │ PASS   │ 0:00:02.978 │
│ actorle-2022-apr-03.txt │ ********        │ ********            │ PASS   │ 0:00:02.867 │
│ actorle-2022-apr-17.txt │ ********        │ ********            │ PASS   │ 0:00:02.936 │
│ actorle-2023-jan-15.txt │ ********        │ ********            │ PASS   │ 0:00:03.362 │
│ actorle-2023-jan-01.txt │ ********        │ ********            │ PASS   │ 0:00:03.031 │
│ actorle-2023-aug-22.txt │ Gabourey Sidibe │ Brahmanandam        │ FAIL   │ 0:00:02.529 │
│ actorle-2022-sep-02.txt │ ********        │ ********            │ PASS   │ 0:00:03.100 │
│ actorle-2022-aug-01.txt │ ********        │ ********            │ PASS   │ 0:00:02.805 │
│ actorle-2022-aug-29.txt │ ********        │ ********            │ PASS   │ 0:00:02.913 │
│ actorle-2022-aug-28.txt │ ********        │ ********            │ PASS   │ 0:00:02.803 │
│ actorle-2023-feb-20.txt │ ********        │ ********            │ PASS   │ 0:00:02.508 │
│ actorle-2022-mar-22.txt │ ********        │ ********            │ PASS   │ 0:00:02.778 │
│ actorle-2022-may-17.txt │ ********        │ ********            │ PASS   │ 0:00:02.806 │
│ actorle-2022-sep-03.txt │ ********        │ ********            │ PASS   │ 0:00:02.676 │
│ actorle-2022-sep-01.txt │ ********        │ ********            │ PASS   │ 0:00:02.685 │
│ actorle-2023-dec-19.txt │ ********        │ ********            │ PASS   │ 0:00:03.184 │
│ actorle-2022-may-01.txt │ ********        │ ********            │ PASS   │ 0:00:02.843 │
│ actorle-2022-mar-20.txt │ ********        │ ********            │ PASS   │ 0:00:02.640 │
│ actorle-2022-jun-18.txt │ ********        │ ********            │ PASS   │ 0:00:02.709 │
│ actorle-2023-oct-22.txt │ ********        │ ********            │ PASS   │ 0:00:03.054 │
│ actorle-2022-mar-19.txt │ ********        │ ********            │ PASS   │ 0:00:02.820 │
│ actorle-2022-jul-31.txt │ ********        │ ********            │ PASS   │ 0:00:02.330 │
│ actorle-2022-mar-30.txt │ ********        │ ********            │ PASS   │ 0:00:02.593 │
│ actorle-2022-mar-18.txt │ ********        │ ********            │ PASS   │ 0:00:02.711 │
│ actorle-2022-sep-05.txt │ ********        │ ********            │ PASS   │ 0:00:02.629 │
│ actorle-2023-feb-18.txt │ ********        │ ********            │ PASS   │ 0:00:02.873 │
│ actorle-2022-aug-04.txt │ ********        │ ********            │ PASS   │ 0:00:02.600 │
│ actorle-2023-feb-25.txt │ ********        │ ********            │ PASS   │ 0:00:02.623 │
│ actorle-2022-mar-27.txt │ ********        │ ********            │ PASS   │ 0:00:03.089 │
│ actorle-2023-oct-20.txt │ ********        │ ********            │ PASS   │ 0:00:03.052 │
│ actorle-2023-dec-07.txt │ ********        │ ********            │ PASS   │ 0:00:03.082 │
│ actorle-2023-feb-28.txt │ ********        │ ********            │ PASS   │ 0:00:03.036 │
│ actorle-2023-feb-01.txt │ ********        │ ********            │ PASS   │ 0:00:02.841 │
│ actorle-2022-mar-17.txt │ ********        │ ********            │ PASS   │ 0:00:02.685 │
│ actorle-2022-jun-06.txt │ ********        │ ********            │ PASS   │ 0:00:03.022 │
│ actorle-2022-sep-21.txt │ ********        │ ********            │ PASS   │ 0:00:03.200 │
│ actorle-2023-feb-07.txt │ ********        │ ********            │ PASS   │ 0:00:03.145 │
│ actorle-2023-apr-07.txt │ ********        │ ********            │ PASS   │ 0:00:03.228 │
│ actorle-2022-may-18.txt │ ********        │ ********            │ PASS   │ 0:00:02.766 │
│ actorle-2022-sep-18.txt │ ********        │ ********            │ PASS   │ 0:00:02.521 │
│ actorle-2022-may-26.txt │ ********        │ ********            │ PASS   │ 0:00:02.743 │
│ actorle-2022-aug-31.txt │ ********        │ ********            │ PASS   │ 0:00:03.029 │
│ actorle-2022-aug-25.txt │ ********        │ ********            │ PASS   │ 0:00:02.588 │
│ actorle-2022-aug-30.txt │ ********        │ ********            │ PASS   │ 0:00:03.366 │
│ actorle-2023-feb-04.txt │ ********        │ ********            │ PASS   │ 0:00:03.035 │
│ actorle-2023-jan-19.txt │ ********        │ ********            │ PASS   │ 0:00:03.029 │
│ actorle-2023-jan-20.txt │ ********        │ ********            │ PASS   │ 0:00:02.989 │
└─────────────────────────┴─────────────────┴─────────────────────┴────────┴─────────────┘
                            2 failed, 51 passed in 0:00:02.991

PYTHONPATH=$PYTHONPATH:. python tests/integration/accuracy-test.py -p  -a  -s  167.73s user 125.39s system 192% cpu 2:31.89 total
```

## After: faster and more accurate
Note that I added a puzzle in this branch too, so the integration accuracy test runs 9 seconds faster __despite__ solving one extra puzzle (which adds at least a couple of seconds), so a like-for-like comparison may be more like 12 seconds faster.
```
time PYTHONPATH=$PYTHONPATH:. python tests/integration/accuracy-test.py -p clues-files -a data/puzzle-actual-answers.csv -s /Users/mickyfitz/workspace/knacktorle/solve-current-actorle-puzzle.sh
                                      Summary
┏━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━┳━━━━━━━━━━━━━┓
┃ Puzzle                  ┃ Expected Answer ┃ Actual Answer ┃ Result ┃ Time        ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━╇━━━━━━━━━━━━━┩
│ actorle-2023-jan-10.txt │ ********        │ ********      │ PASS   │ 0:00:02.815 │
│ actorle-2023-jan-04.txt │ ********        │ ********      │ PASS   │ 0:00:02.688 │
│ actorle-2022-dec-15.txt │ ********        │ ********      │ PASS   │ 0:00:02.758 │
│ actorle-2023-mar-07.txt │ ********        │ ********      │ PASS   │ 0:00:02.406 │
│ actorle-2022-apr-04.txt │ ********        │ ********      │ PASS   │ 0:00:02.561 │
│ actorle-2022-apr-28.txt │ ********        │ ********      │ PASS   │ 0:00:02.490 │
│ actorle-2023-jan-17.txt │ ********        │ ********      │ PASS   │ 0:00:02.576 │
│ actorle-2022-apr-03.txt │ ********        │ ********      │ PASS   │ 0:00:02.621 │
│ actorle-2022-apr-17.txt │ ********        │ ********      │ PASS   │ 0:00:02.569 │
│ actorle-2023-jan-15.txt │ ********        │ ********      │ PASS   │ 0:00:02.919 │
│ actorle-2023-jan-01.txt │ ********        │ ********      │ PASS   │ 0:00:02.554 │
│ actorle-2023-aug-22.txt │ ********        │ ********      │ PASS   │ 0:00:02.375 │
│ actorle-2022-sep-02.txt │ ********        │ ********      │ PASS   │ 0:00:02.731 │
│ actorle-2022-aug-01.txt │ ********        │ ********      │ PASS   │ 0:00:02.526 │
│ actorle-2022-aug-29.txt │ ********        │ ********      │ PASS   │ 0:00:02.671 │
│ actorle-2022-aug-28.txt │ ********        │ ********      │ PASS   │ 0:00:02.605 │
│ actorle-2023-feb-20.txt │ ********        │ ********      │ PASS   │ 0:00:02.405 │
│ actorle-2022-mar-22.txt │ ********        │ ********      │ PASS   │ 0:00:02.530 │
│ actorle-2022-may-17.txt │ ********        │ ********      │ PASS   │ 0:00:02.557 │
│ actorle-2022-sep-03.txt │ ********        │ ********      │ PASS   │ 0:00:02.540 │
│ actorle-2022-sep-01.txt │ ********        │ ********      │ PASS   │ 0:00:02.610 │
│ actorle-2023-dec-19.txt │ ********        │ ********      │ PASS   │ 0:00:02.855 │
│ actorle-2022-may-01.txt │ ********        │ ********      │ PASS   │ 0:00:02.585 │
│ actorle-2022-mar-20.txt │ ********        │ ********      │ PASS   │ 0:00:02.571 │
│ actorle-2022-jun-18.txt │ ********        │ ********      │ PASS   │ 0:00:02.601 │
│ actorle-2023-oct-22.txt │ ********        │ ********      │ PASS   │ 0:00:02.842 │
│ actorle-2022-mar-19.txt │ ********        │ ********      │ PASS   │ 0:00:02.659 │
│ actorle-2022-jul-31.txt │ ********        │ ********      │ PASS   │ 0:00:02.547 │
│ actorle-2022-mar-30.txt │ ********        │ ********      │ PASS   │ 0:00:02.593 │
│ actorle-2022-mar-18.txt │ ********        │ ********      │ PASS   │ 0:00:02.658 │
│ actorle-2023-dec-21.txt │ ********        │ ********      │ PASS   │ 0:00:02.810 │
│ actorle-2022-sep-05.txt │ ********        │ ********      │ PASS   │ 0:00:02.520 │
│ actorle-2023-feb-18.txt │ ********        │ ********      │ PASS   │ 0:00:02.682 │
│ actorle-2022-aug-04.txt │ ********        │ ********      │ PASS   │ 0:00:02.550 │
│ actorle-2023-feb-25.txt │ ********        │ ********      │ PASS   │ 0:00:02.497 │
│ actorle-2022-mar-27.txt │ ********        │ ********      │ PASS   │ 0:00:02.633 │
│ actorle-2023-oct-20.txt │ ********        │ ********      │ PASS   │ 0:00:02.722 │
│ actorle-2023-dec-07.txt │ ********        │ ********      │ PASS   │ 0:00:02.595 │
│ actorle-2023-feb-28.txt │ ********        │ ********      │ PASS   │ 0:00:02.763 │
│ actorle-2023-feb-01.txt │ ********        │ ********      │ PASS   │ 0:00:02.586 │
│ actorle-2022-mar-17.txt │ ********        │ ********      │ PASS   │ 0:00:02.495 │
│ actorle-2022-jun-06.txt │ ********        │ ********      │ PASS   │ 0:00:02.590 │
│ actorle-2022-sep-21.txt │ ********        │ ********      │ PASS   │ 0:00:02.768 │
│ actorle-2023-feb-07.txt │ ********        │ ********      │ PASS   │ 0:00:02.737 │
│ actorle-2023-apr-07.txt │ ********        │ ********      │ PASS   │ 0:00:02.779 │
│ actorle-2022-may-18.txt │ ********        │ ********      │ PASS   │ 0:00:02.662 │
│ actorle-2022-sep-18.txt │ ********        │ ********      │ PASS   │ 0:00:02.386 │
│ actorle-2022-may-26.txt │ ********        │ ********      │ PASS   │ 0:00:02.561 │
│ actorle-2022-aug-31.txt │ ********        │ ********      │ PASS   │ 0:00:02.765 │
│ actorle-2022-aug-25.txt │ ********        │ ********      │ PASS   │ 0:00:02.479 │
│ actorle-2022-aug-30.txt │ ********        │ ********      │ PASS   │ 0:00:02.819 │
│ actorle-2023-feb-04.txt │ ********        │ ********      │ PASS   │ 0:00:02.862 │
│ actorle-2023-jan-19.txt │ ********        │ ********      │ PASS   │ 0:00:02.957 │
│ actorle-2023-jan-20.txt │ ********        │ ********      │ PASS   │ 0:00:02.855 │
└─────────────────────────┴─────────────────┴───────────────┴────────┴─────────────┘
                         0 failed, 54 passed in 0:00:02.857

PYTHONPATH=$PYTHONPATH:. python tests/integration/accuracy-test.py -p  -a  -s  154.38s user 141.72s system 207% cpu 2:22.69 total
```